+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The change log is available [on GitHub][2].
 * [#43](https://github.com/kowainik/typerep-map/issues/43):
   Implement `IsList` instance for `TypeRepMap`.
   Add `hoistA` function.
+* [#39](https://github.com/kowainik/typerep-map/issues/39):
+  Implement `hoistWithKey` function.
+  Add `map` function for `TMap`.
 * Drop support for `ghc-8.0.2`.
 
 # 0.1.0